:sparkles:  add http gateway and change serviceid type from ushort to int32
新增http网关功能，修复服务端拦截器的类型错误，新增http网关的示例，修改序列化方案中必须固定CodecType字段

+++ src/DotBPE.Gateway/DefaultImpl/HttpRequestData.cs
@@ -5,7 +5,7 @@ namespace DotBPE.Gateway
 {
     public class HttpRequestData:IHttpRequestData
     {
-        public ushort ServiceId { get; set; }
+        public int ServiceId { get; set; }
         public ushort MessageId { get; set; }
         public IDictionary<string, string> QueryOrFormData { get;  } = new Dictionary<string, string>();
         public string RawBody { get; set; }
